slack_notion_integration/requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
typing-extensions==4.8.0

# Slack integration
slack-bolt==1.18.0
slack-sdk==3.23.0

# Notion integration
notion-client==2.0.0
notion-sdk-py==0.1.3

# LangGraph and related
langgraph==0.0.19
archon==0.0.5
langchain==0.0.340

# Language Models
openai==1.3.7
anthropic==0.5.0
together==0.2.3

# Vector Database
pinecone-client==2.2.4
redis==4.6.0
psycopg2-binary==2.9.9
pgvector==0.2.3
sqlalchemy==2.0.23
alembic==1.12.1

# Context Management
tiktoken==0.5.1
numpy==1.24.3
sentence-transformers==2.2.2

# Async and Concurrency
aiohttp==3.8.6
asyncio==3.4.3
aiocache==0.12.2
tenacity==8.2.3

# Utility
requests==2.31.0
python-dateutil==2.8.2
uuid==1.30
tqdm==4.66.1
rich==13.6.0
loguru==0.7.2

# Web server
fastapi==0.103.1
uvicorn==0.23.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0